# Specialize interleave string ~2-3x faster