# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Part of #2864

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

```
interleave str(20, 0.0) 100 [0..100, 100..230, 450..1000]
                        time:   [1.0792 µs 1.0795 µs 1.0799 µs]
                        change: [-60.193% -60.122% -60.019%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

interleave str(20, 0.0) 400 [0..100, 100..230, 450..1000]
                        time:   [3.2179 µs 3.2187 µs 3.2197 µs]
                        change: [-60.739% -60.622% -60.497%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) low mild
  2 (2.00%) high mild
  5 (5.00%) high severe

interleave str(20, 0.0) 1024 [0..100, 100..230, 450..1000]
                        time:   [7.4057 µs 7.4079 µs 7.4101 µs]
                        change: [-60.606% -60.542% -60.442%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

interleave str(20, 0.0) 1024 [0..100, 100..230, 450..1000, 0..1000]
                        time:   [7.4336 µs 7.4369 µs 7.4403 µs]
                        change: [-58.701% -58.630% -58.525%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe

interleave str(20, 0.5) 100 [0..100, 100..230, 450..1000]
                        time:   [1.6810 µs 1.6815 µs 1.6821 µs]
                        change: [-58.572% -58.434% -58.308%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  2 (2.00%) high severe

interleave str(20, 0.5) 400 [0..100, 100..230, 450..1000]
                        time:   [5.2158 µs 5.2246 µs 5.2341 µs]
                        change: [-64.090% -63.972% -63.864%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

interleave str(20, 0.5) 1024 [0..100, 100..230, 450..1000]
                        time:   [16.993 µs 17.013 µs 17.032 µs]
                        change: [-54.202% -54.077% -53.991%] (p = 0.00 < 0.05)
                        Performance has improved.

interleave str(20, 0.5) 1024 [0..100, 100..230, 450..1000, 0..1000]
                        time:   [17.424 µs 17.443 µs 17.464 µs]
                        change: [-52.353% -52.256% -52.128%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

Adds a specialized implementation for string arrays

# Are there any user-facing changes?

No

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->