@tustvold 2022-10-27
With #2947 this will be generalizable to also cover byte arrays